requirements.txt
aiohttp==3.6.3
appnope==0.1.0
argon2-cffi @ file:///opt/concourse/worker/volumes/live/b46b812d-eebc-4040-75c0-7fc24d9bfd54/volume/argon2-cffi_1596828479252/work
asgiref==3.3.1
astroid @ file:///opt/concourse/worker/volumes/live/76ddd801-8c2e-42ae-78fe-7e7640c9eb32/volume/astroid_1592495930711/work
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
bs4==0.0.1
certifi==2020.12.5
cffi @ file:///opt/concourse/worker/volumes/live/e2eb91c2-780d-4f74-545d-c2827857ffe3/volume/cffi_1600699193708/work
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
constantly==15.1.0
crayons==0.4.0
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
elasticsearch==5.5.3
entrypoints==0.3
eventkit==0.8.6
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
finviz==1.3.4
gunicorn==20.0.4
hyperlink==20.0.1
ib-insync==0.9.63
idna==2.10
idna-ssl==1.1.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
incremental==17.5.0
ipykernel @ file:///opt/concourse/worker/volumes/live/43eaf185-e9cd-4e78-5afa-c19accf1339d/volume/ipykernel_1596206683810/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/8c453bb6-500b-4747-778c-a236ae4a574d/volume/ipython_1593447395779/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
isort @ file:///tmp/build/80754af9/isort_1612243941122/work
itemadapter==0.2.0
itemloaders==1.0.4
jedi @ file:///opt/concourse/worker/volumes/live/565dbe5a-443f-465a-7ff2-b61f4a5f081d/volume/jedi_1598371617578/work
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver==1.2.0
lazy-object-proxy @ file:///opt/concourse/worker/volumes/live/0ac78eb4-a09b-44ae-4f3a-0b00b7034ce6/volume/lazy-object-proxy_1607707349885/work
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mpl-finance==0.10.1
mplfinance==0.12.7a0
multidict==4.7.6
multitasking==0.0.9
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///opt/concourse/worker/volumes/live/83354520-8b79-4a5c-771c-53c9f422edfc/volume/nbconvert_1601914843271/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.5
notebook @ file:///opt/concourse/worker/volumes/live/cdd50f32-e012-4ec4-5f72-cf8d8a902618/volume/notebook_1601501566228/work
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==3.4.5.20
opencv-python-headless==4.4.0.46
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
Protego==0.1.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyDispatcher==2.0.5
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
PyHamcrest==2.0.2
pylint @ file:///opt/concourse/worker/volumes/live/7cea4525-4763-4343-4314-b6c7c20dbe1f/volume/pylint_1598624000729/work
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent @ file:///opt/concourse/worker/volumes/live/d7ed6755-1956-4a16-5499-c03638db758d/volume/pyrsistent_1600141722278/work
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
queuelib==1.5.0
readability-lxml==0.8.1
regex==2020.11.13
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.3
Scrapy==2.4.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
statsmodels==0.12.0
TA-Lib==0.4.19
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tornado==4.5.2
tqdm==4.54.1
traitlets==4.3.3
tweepy==3.9.0
Twisted==20.3.0
twitter==1.18.0
typed-ast @ file:///opt/concourse/worker/volumes/live/85c87157-84ad-443c-5815-d7450a5595c1/volume/typed-ast_1610484564230/work
typing-extensions==3.7.4.3
urllib3==1.25.11
user-agent==0.1.9
vaderSentiment==3.3.2
w3lib==1.22.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webdriver-manager==3.2.2
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt @ file:///opt/concourse/worker/volumes/live/4f9b85ff-0e2b-45ea-5ced-59ae4d0157bf/volume/wrapt_1607712503374/work
XlsxWriter==1.3.7
yarl==1.5.1
yfinance==0.1.55
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work
zope.interface==5.2.0
